fix(tcbs): correct DTO types and cleanup examples
- BankAccount.Authorized, StockPurchasingPower rates -> decimal.Decimal
- PageNo/PageSize -> int64 (was float64)
- Add NumDayMarginCall to MarginAccountInformationRiskPolicy
- GetMarginAccountInformationResponse is a single object, not slice
- StockMatchingDetail.TimeExec -> string, TradeID -> decimal.Decimal
- Use setOptional for optional derivative order query params
- Fix query param toDate -> todate (lowercase)
- Swap examples to env.String with defaults, fix env var names
- Add endpoint debug print in do() helper

+++ brokers/tcbs/native/dto/derivative_asset_operations.go
@@ -18,8 +18,8 @@ type GetClosedDerivativePositionsRequest struct {
 	AccountID    string
 	SubAccountID string
 	Symbol       string
-	PageNo       float64
-	PageSize     float64
+	PageNo       int64
+	PageSize     int64
 }
 
 type GetClosedDerivativePositionsResponse struct {
